requirements.txt
django==3.2.2
djangorestframework==3.11.0
mysqlclient==2.0.3
